move bokeh server communication to background page
