+++ distributed/bokeh/application.py
@@ -17,7 +17,7 @@ from ..config import config
 
 dirname = os.path.dirname(distributed.__file__)
 paths = [os.path.join(dirname, 'bokeh', name)
-         for name in ['status', 'tasks', 'workers']]
+         for name in ['background', 'status', 'tasks', 'workers']]
 
 logger = logging.getLogger(__file__)
