+++ setup.py
@@ -25,6 +25,7 @@ setup(name='distributed',
                 'distributed.cli',
                 'distributed.diagnostics',
                 'distributed.bokeh',
+                'distributed.bokeh.background',
                 'distributed.bokeh.status',
                 'distributed.bokeh.tasks',
                 'distributed.bokeh.workers',
